method syntax

+++ rectangles/src/main.rs
@@ -5,7 +5,7 @@ struct Rectangle {
 }
 
 impl Rectangle {
-    fn area(&self) -> u32 {
+    fn area(&self) -> u64 {
         self.width * self.height
     }
